Renamed moxfield client to api so we can build a client layer ontop of it
